# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API барои системаи тиҷоратии электронӣ",
        "description": "Санади API барои ҳама контроллерҳо",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Воридшавӣ ба система",
                "description": "Истифодабаранда бо email ва password ворид мешавад",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почтаи электронӣ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "user@example.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Рамз",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password",
                            "example": "12345678"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Воридшавӣ бомуваффақият анҷом ёфт",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefg123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Хатогии валидация"
                    },
                    "401": {
                        "description": "Email ё password нодуруст аст"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Берун рафтан",
                "description": "Нест кардани token-и корбар",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Берун рафтани муваффақ"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ба қайд гирифтан",
                "description": "Эҷод кардани корбари нав",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Ном",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ali"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Насаб",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ahmadov"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Почтаи электронӣ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "user@example.com"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Рақами телефон",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "+992900000000"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Рамз",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345678"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Тасдиқи рамз",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345678"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ба қайд гирифтан муваффақ шуд",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefg123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Хатогии валидация"
                    }
                }
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Паролро иваз кардан",
                "description": "Тағйир додани пароли корбар",
                "operationId": "540baf9b2324b156be71c57911de6714",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|abcdefg123456"
                        }
                    },
                    {
                        "name": "current_password",
                        "in": "query",
                        "description": "Пароли ҳозира",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345678"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "Пароли нав",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "87654321"
                        }
                    },
                    {
                        "name": "new_password_confirmation",
                        "in": "query",
                        "description": "Тасдиқи пароли нав",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "87654321"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Парол бомуваффақият иваз шуд"
                    },
                    "422": {
                        "description": "Хатогии валидация"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Маълумоти корбар",
                "description": "Гирифтани маълумоти корбари ҷорӣ",
                "operationId": "7c1cf647d76fad4b3eb1549fa4fe2f4c",
                "responses": {
                    "200": {
                        "description": "Муваффақ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Ali"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Ahmadov"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "user@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}